Work on the CGO menu
The Custom Game Options menu is now accessible via the normal Options Menu, for now.

Fixed a few bugs in how menu items' values are displayed. Unfortunately, the menu as a whole is... well,...

GBA graphics are hard.

+++ include/lu/custom_game_options_menu/menu_item.h
@@ -53,7 +53,7 @@ u16  GetOptionValue(const struct CGOptionMenuItem* item);
 void SetOptionValue(const struct CGOptionMenuItem* item, u16 value);
 
 const u8* GetOptionValueName(const struct CGOptionMenuItem* item, u16 value);
-u8        GetOptionValueCount(const struct CGOptionMenuItem* item);
+u16       GetOptionValueCount(const struct CGOptionMenuItem* item);
 u16       GetOptionMinValue(const struct CGOptionMenuItem* item);
 
 void CycleOptionSelectedValue(const struct CGOptionMenuItem* item, s8 by);
